Explicit error if _DEFAULT_LOCALE_PATH is None
Try to use __file__ if possible but do not crash if not.

Instead, when `activate` is used and no explicit path is given, raise an
exception explaining why.

A better alternative could be use `resource_path`
(https://docs.python.org/3/library/importlib.html#importlib.abc.ResourceReader)
to find the directory of 'locale'.

Unfortunately Python (at least 3.7) only allows to access to files
and not folders so we cannot access to 'locale' (a folder)

Closes #81