🐑 Add seperate typing hints for humanize
Adds a collection of `.pyi` files that describe the expected type
signatures of the publicly available `humanize` functions.

A `MANIFEST.in` file has been added to include the new `.pyi` files and
the `py.typed` file to indicate to mypy that this package supports
typing.

mypy has been added to the pre-commit configuration to check the stubs.

Some of the signatures are slightly optimistic, since depending on error
conditions, different types are returned.