🐑 Remove stubs, add typing to functions
It is better to include typing hints directly in the functions that
they're describing; which allows mypy to actually check the functions,
both in their execution, and that they are actually returning the
appropriate values.

The major change here is that previously a number of the functions, if
they encountered a conversion error, would return the value unchanged.
However, that would require just declaring the return type as Any, which
although correct, isn't particularly useful.

So instead, any failed, unconvertable, inconsistent values will just
have `str` called on them, and the output will returned. The function
will always return a string, although sometimes it won't be in the
format you expect.

In addition, some internal variables inside the function have been
rearranged and moved around, to avoid changing the type of an already
defined variable.