Only show exported functions in docs for i18n module
Prior to this change, all of the public functions in the 18n module were
dhown in the public documentation. However, these were not all available
from the root package.

This change includes only the exported functions in the documentation.

The "private" functions could have been hidden by prefixing them with
`_`, but this would have constituted a rename and might be a breaking
change. Therefor we take the approach of hiding them from the
documentation rather than making them proper private functions.