Make functions private and deprecate public aliases in time module
Prior to this change, the `Unit` enum and the `abs_timedelta` and
`date_and_delta` functions were listed as public functions in the
documentation but were not available from the root package:

    >>> import humanize
    >>> humanize.Unit
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    AttributeError: module 'humanize' has no attribute 'Unit'
    >>> humanize.abs_timedelta
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    AttributeError: module 'humanize' has no attribute 'abs_timedelta'

This change makes these members private.

In order to preserve backwards compatibility after this change, we
provide aliases for these members which emit deprecation warnings.