Make functions private and deprecate public aliases in i18n module
Prior to this change, the `Unit` enum and the `abs_timedelta` and
`date_and_delta` functions were listed as public functions in the
documentation but were not available from the root package:

    >>> import humanize
    >>> humanize.gettext
    Traceback (most recent call last):
        File "<stdin>", line 1, in <module>
    AttributeError: module 'humanize' has no attribute 'gettext'
    >>> humanize.ngettext
    Traceback (most recent call last):
        File "<stdin>", line 1, in <module>
    AttributeError: module 'humanize' has no attribute 'ngettext'

This change makes these members private.

In order to preserve backwards compatibility after this change, we
provide aliases for these members which emit deprecation warnings.